requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
joblib
tensorflow
comet_ml
dvc